2026-g-1/lec-2/pyproject.toml
[tool.poetry]
name = "lec-1"
version = "0.1.0"
description = ""
authors = ["jaba.kochashvili.1 <jaba.kochashvili.1@btu.edu.ge>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.42.68"
python-dotenv = "^1.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
